Use the unify_stap_output script to simplify comparisons of different binaries
"stap -L ..." often includes full paths in the output.  The different
paths in the output for different builds of the same source code with
various flags makes comparison and analysis more difficult.  The
unif_stap_output script removes most of the path to allow the diff
command to be used between the stap output from different binaries.