Factor out the driver to iterate over the variations of compiler options
Wanted to make it analysis scripts smaller and minimize duplication in
the code.  Put the driver in a separate file that is included by
various scripts.  The gen_rpm_variants scripts does all the setup
build directories. The analysis scripts only read and do not modify
the build directories. This is why the set_debug_links was moved from
the analyze phase into the build_rpm in gen_rpm_variants script.