Disable spack v0.13.0 auto detection of target
In an attempt to generate tuned code spack v0.13.0 and newer will
identify the microarchitecture and use that at the target.  We don't
want spack picking different tuning based on the specifics of the
build system.  The gen_spack_variants code needs to explicitly select
the generic architecture as a target to prevent spack from auto tuning
for a random build environment.