Work around disjoint code for function being lumped together
Dyninst lumps symbol table entries for disjoint sections of code
together, but it treats the disjoint sections of code as separate
functions, so it is when searching for the location lists for a
function it gets entries for other disjoint regions.  This check is to
see whether any of the range overlaps with the function currently
being examined.