Add cse.C proof of concept code
This code analyzes the location lists for functions and outputs areas in the
code where a particular register maps back to multiple variables.  This could
be an issue for live patching.  Changing one variable could unintentially
change the value for the other variable.

This code only analyzes values stored in registers.  The assumption is
that values in memory are actual variables.  However, this might miss
the areas where a value in a register is spilled into temporary
storage on the stack.