readonly code prints locations where register variables are read-only
readonly now examines the location information to determine whether
the register holds a variable listed in a location list.  The liveness
analysis is done if there is a variable in the register when then
instruction is executed.  If the register is dead (changes to register
value at this address have no effect), a message is printed out.

The code also prints a diagnostic dump of the location list
information before doing the analysis to aid in verifying the readonly
code is working properly.  This debugging code will be disabled later.

are detm