Add options to readonly.
Added code to do simple option parsing in readonly.  Currently the
only option handled is one to turn on optional generation of debugging
output.  The code has been added to make it easy to split the the
option handling and some of the other common code into a library to
reduce the size of the analysis tools.