Add filtering logic and option
This patch provides a mechanism to store a information for location of
interest in the code that can be retrieved based on location by the
interesting_loc function.  If there are no filters the mechanism just
stores the function name for area of a function, which is similar to
the previous default behavior.  The "-f" option is recognized as a
valid option, but hasn't been implemented.  It will allow selection of
function and line numbers to mark as points of interest.