Add regular expressions to match common systemtap probes for '-f' arguments
The matching now allows common systemtap probe to be passed into readonly.
The readonly program parses the following common probes:

   process("PATH").function("NAME").call
   process("PATH").function("NAME").return
   process("PATH").statement("*@FILE.c:123")
   process("PATH").function("NAME").inline
   process("PATH").function("NAME")

Currently, the code does not mark locations of inlined functions, so
the last two probes in the list will miss a lot of locations of
interest.  Wildcards are allowd for function names.  However,
wildcards are not yet implmented for statement file names or lines.