Analyze all possible instructions in a function.
The Linux kernel tracepoint code ends up being put at the end
of functions with no visible means of executing it.  There are
no-op instruction in the locations that will be patched with
jumps to the tracepoint code when it is activated.  The downside
of this type of code is that Dyninst's basic block analysis missed
those snippets of code related to the inactive tracepoints.  This
results in a huge number of warnings about location ranges
not being on valid boundaries from locationrangecheck.

The location range check doesn't need the basic block analysis
information.  The code now just decodes the size of each instruction
for the space that the function occupies.  This allows
locationrangecheck to analyze the location ranges for the inactive
tracepoint code.