Update const_parameters to provide more data and handle location list consts
The const_parameter script now provides percentages on the number of
function and inlined functions that have constant parameters and the
percentage of parameters that are constants for the different flavor
of functions.

There are several ways that the compiler may encode constants in the
DWARF information.  Initially, the script just looked for the
DW_AT_const_value attribute and this missed information in location
lists.  The script makes the assumption that the first element in the
location list covers the entry to the function.